Modal Styled More or Less
I really should have been doing media queries this whole time...

EditNote.js --- Changed it to show the note it edited instead of going back to the list of notes.

ViewNote.js --- Modal styling.

Tried installing a differnt modal that was supposedly simpler and allowed for className.  Ended up sticking with the original.