Basic MVP Feature Functionality
Imported Redux and created a store with the notes we want to display.
Connected and mapped state to props so that each component can access the
data.

List view displays all notes in the notes object within the state.

Sidebar has two buttons. View Notes button links back to mainpage. Create
Note button links to a CreateNotes component.

Clicking an individual note routes to a Notes component that displays
the content of that specific note by using the match.params.id in the props
provided by our router.

Within each note is an Edit and Delete button which will route to their
respective components.

Currently working on the Delete modal. And then hopefully begin
styling the components.