Search Working sort of
Search will work, but only for more or less exact terms.  For instance, it will find rick's if I type in rick's, but not if I search for rick.

I prettified every file... so this commit message will be pretty worthless.  Per the usual, I added a component/action/reducer...