Create Note progress and started on Delete Modal
Action and Reducer for creating a new note is complete.
CreateNewNote component accepts Note Title and Content and saves
to state and redirects back to homepage upon clicking "Save."

Next features to be worked on are Edit and Delete. Then hopefully making
it all look nice with styling.

Implemented a basic Delete Modal for now.