However Unnecessary, Chronological Sort Working
It occurred to me that the notes will always be chronologically sorted...  Regardless, I set on
e of the notes in initial state to an earlier date and it was became first in the list.

SortOverView.js --- added function to sort chronologically.

reducers/index.js --- Modified the date of the 3rd note titled 'Premise'.