Sort Edit/Delete broken still
Upon fixing edit/delete in regards to an edited note, I noticed it is still broken for when you are viewing a sorted list.

The way I have sortedlist set up, it doesn't actually pass the note to Note component, but rather an array.  Long story short, after messing with this for an hour, I've come to realize the scope of fixing this is expansive.

I instead resort back to my old ways of not having edit/delete on a sorted list for the time being.