Login Working to Some Degree
At present, Every Component works that doesn't involve creating/editing.  All that should take is modifying the reducers I believe.

As far as modifications on components, I changed state.notes to state.currentUserNotes, and obviously, added currentUserNotes to initialState that is updated on login.

At present, how this is working, is the initial state has users with a username,password,notes array.  When logged in, their notes are stored in currentUserNotes array.  When everything is working, I will have a logout function that replaces that with an empty array.  This may or may not be necessary.