Add ability to reset search and filteredArray
Resets the search field and filteredArray when page is refreshed or navigation to other components occurs.

This ensures that when you navigate back to view notes the notes are displayed again.

It might be a good idea to save the state of the search field and keep it filtered, but it introduces other problems with state.

Also removed the service worker as you need https to use it.