working on display
I keep fighting having to divide this project into components but I think I'm going to have to. As long as I can work on all of my components at the same time, I'll be able to do it. Here, I created some components for displaying the home page and started working on the main part.