Added additional folders
and files. To better represent what the app will be when complete. Also added Bootstrap and Reactstrap (the correct way) and will be removing the URL CDN shortly after this commit. I thought long and hard about it and I decided that I can't make this app without utilizing redux from the beginning. It would be disappointing to myself and my instructors. I can do better than that and this deserves my best.