Redid all my styling
decided it was silly to have it all in separate files since a lot of the stuff would be reused and I could just use the same class names or easily see what I did for a different form/view and modify it all in one document. Finished ListView, CreateView, and NoteView - now I'm working on EditView.