MVP with styling done
I finally figured out how to get rid of the text decoration on hover in my list view. The only thing I couldn't get was removing the border radius from the delete modal, but that is not a big thing so I think I'm just going to let it go. Going to look at some of the stretch items to see if any of them are doable in a day.