EXTRA - updated Edit View
It was bothering me that if you clicked on edit, it would take you to a screen with blank inputs instead of showing you what you were trying to edit. I changed it so that the inputs would contain the existing value of the notecard, so if you just want to fix a typo or add a word or something, you don't need to rewrite the whole thing.