Front End Project Week - redux
Set up and wired containers for Redux. Connected them into the app flow. Everything still working as expected so far.
Was getting a warning about using "exact" on a Router Link and not explicitly setting it to a boolean. Suspect something to do with Styling it with styled-components. Explicitly set it to "true" for the time being, which cleared the warning.
Got most of the reducers wrote out, except for GOT_SINGLE_NOTE, which I need to check on its return data, then adjust initial state and action accordingly.
Some minor styling tweaks